This is a bunch of changes including: * Turning on x64 toolchain host support * Support for Visual Studio 2017 (15.9) including C++ compliance changes. * Eliminating MFC headers * Turning on /GS stack checks in all binary targets"

+++ omaha/crashhandler/crash_dump_util.cc
@@ -73,7 +73,8 @@ HRESULT GetEnvironmentVariableToBuffer(const TCHAR* name,
                                    env_value_char.GetLength(),
                                    static_cast<char*>(value_out),
                                    static_cast<int>(value_buffer_size));
-  if (result_size < 0 || value_buffer_size != result_size) {
+  if (result_size < 0 ||
+      value_buffer_size != static_cast<size_t>(result_size)) {
     CORE_LOG(LE, (_T("[Failed to decode value][%s=%s]"), name, env_value));
     return HRESULT_FROM_WIN32(ERROR_INSUFFICIENT_BUFFER);
   }
